+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# [3.33.0](https://github.com/Automattic/newspack-newsletters/compare/v3.32.0...v3.33.0) (2026-05-18)
+
+
+### Bug Fixes
+
+* **block-theme:** override theme.json styles in the newsletter editor ([#2078](https://github.com/Automattic/newspack-newsletters/issues/2078)) ([49c857c](https://github.com/Automattic/newspack-newsletters/commit/49c857c11cf11da39c7e11cca830edb9e77c84ab))
+* ensure editor preview and front end match for default subscription block style ([#2107](https://github.com/Automattic/newspack-newsletters/issues/2107)) ([9624edc](https://github.com/Automattic/newspack-newsletters/commit/9624edc3ee1d0a2bd32898fa9a2a2c8358e30587))
+* hide font family option from newsletter ([#2099](https://github.com/Automattic/newspack-newsletters/issues/2099)) ([3a928a5](https://github.com/Automattic/newspack-newsletters/commit/3a928a5c4a1a8988ea40b68083f5495e735cb565))
+* skip and remove malformed subscription intents ([#2102](https://github.com/Automattic/newspack-newsletters/issues/2102)) ([c0417c6](https://github.com/Automattic/newspack-newsletters/commit/c0417c68d33e0fcc1f98f1e0f59e1bc11d85e16a))
+
+
+### Features
+
+* **esp:** enrich contact field schema for integrations ([#2082](https://github.com/Automattic/newspack-newsletters/issues/2082)) ([3434e75](https://github.com/Automattic/newspack-newsletters/commit/3434e754817c33a3a6fb8171cf28d73c0b39627e))
+* improve newsletter editor layout and post-inserter UX ([#2077](https://github.com/Automattic/newspack-newsletters/issues/2077)) ([d9927cc](https://github.com/Automattic/newspack-newsletters/commit/d9927cca85c0100d705908fc36de9f384656aa57))
+
 # [3.32.0](https://github.com/Automattic/newspack-newsletters/compare/v3.31.1...v3.32.0) (2026-05-04)